requirements.txt
streamlit>=1.40.0,<2.0.0
numpy>=1.21.0
pillow>=9.0.0,<11.0.0
pandas>=1.5.0,<3.0.0
matplotlib>=3.5.0,<4.0.0
plotly>=5.0.0,<6.0.0
scikit-learn>=1.1.0,<2.0.0